requirements.txt
apify-client==1.10.0
python-dotenv
fastapi
pinecone
pinecone[asyncio]
langchain-ollama
langchain-text-splitters
langchain-community
langgraph